doc-anonymizer-v2/requirements.txt
transformers>=4.50
torch
openpyxl
python-docx
python-pptx
pdfplumber
